+++ Carbon.Environment/Functions/Set-CEnvVariable.ps1
@@ -1,75 +1,78 @@
 
-function Set-CEnvironmentVariable
+function Set-CEnvVariable
 {
     <#
     .SYNOPSIS
     Creates or sets an environment variable.
-    
+
     .DESCRIPTION
-    Uses the .NET [Environment class](http://msdn.microsoft.com/en-us/library/z8te35sa) to create or set an environment variable in the Process, User, or Machine scopes.
-    
-    Changes to environment variables in the User and Machine scope are not picked up by running processes.  Any running processes that use this environment variable should be restarted.
+    Uses the .NET [Environment class](http://msdn.microsoft.com/en-us/library/z8te35sa) to create or set an environment
+    variable in the Process, User, or Machine scopes.
 
-    Beginning with Carbon 2.3.0, you can set an environment variable for a specific user by specifying the `-ForUser` switch and passing the user's credentials with the `-Credential` parameter. This will run a PowerShell process as that user in order to set the environment variable.
+    Changes to environment variables in the User and Machine scope are not picked up by running processes.  Any running
+    processes that use this environment variable should be restarted.
 
-    Normally, you have to restart your PowerShell session/process to see the variable in the `env:` drive. Use the `-Force` switch to also add the variable to the `env:` drive. This functionality was added in Carbon 2.3.0.
-    
+    To set an environment variable for a specific user, pass that user's credentials to the `-Credential` parameter.
+    This will run a PowerShell process as that user in order to set the environment variable.
+
+    Normally, you have to restart your PowerShell session/process to see the variable in the `env:` drive. Use the
+    `-Force` switch to also add the variable to the `env:` drive.
+
     .LINK
-    Carbon_EnvironmentVariable
-
-    .LINK
-    Remove-CEnvironmentVariable
+    Remove-CEnvVariable
 
     .LINK
     http://msdn.microsoft.com/en-us/library/z8te35sa
 
     .EXAMPLE
-    Set-CEnvironmentVariable -Name 'MyEnvironmentVariable' -Value 'Value1' -ForProcess
-    
-    Creates the `MyEnvironmentVariable` with an initial value of `Value1` in the process scope, i.e. the variable is only accessible in the current process.
-    
-    .EXAMPLE
-    Set-CEnvironmentVariable -Name 'MyEnvironmentVariable' -Value 'Value1' -ForComputer
-    
-    Creates the `MyEnvironmentVariable` with an initial value of `Value1` in the machine scope, i.e. the variable is accessible in all newly launched processes.
-    
-    .EXAMPLE
-    Set-CEnvironmentVariable -Name 'SomeUsersEnvironmentVariable' -Value 'SomeValue' -ForUser -Credential $userCreds
+    Set-CEnvVariable -Name 'MyEnvironmentVariable' -Value 'Value1' -ForProcess
 
-    Demonstrates that you can set a user-level environment variable for another user by passing its credentials to the `Credential` parameter. Runs a separate PowerShell process as that user to set the environment variable.
+    Creates the `MyEnvironmentVariable` with an initial value of `Value1` in the process scope, i.e. the variable is
+    only accessible in the current process.
+
+    .EXAMPLE
+    Set-CEnvVariable -Name 'MyEnvironmentVariable' -Value 'Value1' -ForComputer
+
+    Creates the `MyEnvironmentVariable` with an initial value of `Value1` in the machine scope, i.e. the variable is
+    accessible in all newly launched processes.
+
+    .EXAMPLE
+    Set-CEnvVariable -Name 'SomeUsersEnvironmentVariable' -Value 'SomeValue' -ForUser -Credential $userCreds
+
+    Demonstrates that you can set a user-level environment variable for another user by passing its credentials to the
+    `Credential` parameter. Runs a separate PowerShell process as that user to set the environment variable.
     #>
-    [CmdletBinding(SupportsShouldProcess=$true)]
+    [CmdletBinding(SupportsShouldProcess)]
     param(
-        [Parameter(Mandatory=$true)]
         # The name of environment variable to add/set.
-        [string]$Name,
-        
-        [Parameter(Mandatory=$true)]
+        [Parameter(Mandatory)]
+        [String] $Name,
+
         # The environment variable's value.
-        [string]$Value,
-        
-        [Parameter(ParameterSetName='ForCurrentUser')]
+        [Parameter(Mandatory)]
+        [String] $Value,
+
         # Sets the environment variable for the current computer.
-        [Switch]$ForComputer,
-
         [Parameter(ParameterSetName='ForCurrentUser')]
-        [Parameter(Mandatory=$true,ParameterSetName='ForSpecificUser')]
+        [switch] $ForComputer,
+
         # Sets the environment variable for the current user.
-        [Switch]$ForUser,
-        
         [Parameter(ParameterSetName='ForCurrentUser')]
+        [Parameter(Mandatory, ParameterSetName='ForSpecificUser')]
+        [switch] $ForUser,
+
         # Sets the environment variable for the current process.
-        [Switch]$ForProcess,
-
         [Parameter(ParameterSetName='ForCurrentUser')]
-        # Set the variable in the current PowerShell session's `env:` drive, too. Normally, you have to restart your session to see the variable in the `env:` drive.
-        #
-        # This parameter was added in Carbon 2.3.0.
-        [Switch]$Force,
+        [switch] $ForProcess,
 
-        [Parameter(Mandatory=$true,ParameterSetName='ForSpecificUser')]
+        # Set the variable in the current PowerShell session's `env:` drive, too. Normally, you have to restart your
+        # session to see the variable in the `env:` drive.
+        [Parameter(ParameterSetName='ForCurrentUser')]
+        [switch] $Force,
+
+        [Parameter(Mandatory,ParameterSetName='ForSpecificUser')]
         # Set an environment variable for a specific user.
-        [pscredential]$Credential
+        [pscredential] $Credential
     )
 
     Set-StrictMode -Version 'Latest'
@@ -80,12 +83,12 @@ function Set-CEnvironmentVariable
         $parameters = $PSBoundParameters
         $parameters.Remove('Credential')
         $job = Start-Job -ScriptBlock {
-            Import-Module -Name (Join-Path -path $using:carbonRoot -ChildPath 'Carbon.psd1' -Resolve)
+            Import-Module -Name (Join-Path -path $using:moduleDirPath -ChildPath 'Carbon.Environment.psm1' -Resolve)
             $VerbosePreference = $using:VerbosePreference
             $ErrorActionPreference = $using:ErrorActionPreference
             $DebugPreference = $using:DebugPreference
             $WhatIfPreference = $using:WhatIfPreference
-            Set-CEnvironmentVariable @using:parameters
+            Set-CEnvVariable @using:parameters
         } -Credential $Credential
         $job | Wait-Job | Receive-Job
         $job | Remove-Job -Force -ErrorAction Ignore
@@ -108,13 +111,13 @@ function Set-CEnvironmentVariable
             {
                 [EnvironmentVariableTarget]::User
             }
-            
+
             if( $Force -or $ForProcess )
             {
                 [EnvironmentVariableTarget]::Process
             }
-        } | 
+        } |
         Where-Object { $PSCmdlet.ShouldProcess( "$_-level environment variable '$Name'", "set") } |
-        ForEach-Object { [Environment]::SetEnvironmentVariable( $Name, $Value, $_ ) }    
+        ForEach-Object { [Environment]::SetEnvironmentVariable( $Name, $Value, $_ ) }
 }