+++ Tests/Remove-CEnvVariable.Tests.ps1
@@ -2,13 +2,22 @@
 #Requires -Version 5.1
 Set-StrictMode -Version 'Latest'
 
+BeforeDiscovery {
+    Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath '..\PSModules\Carbon.Accounts' -Resolve) `
+                  -Function @('Test-CRunAsElevated') `
+                  -Prefix 'T' `
+                  -Verbose:$false
+
+}
+
 BeforeAll {
     Set-StrictMode -Version 'Latest'
 
-    & (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-CarbonTest.ps1' -Resolve)
+    Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath '..\Carbon.Environment' -Resolve) -Verbose:$false
 
     $script:varName = ''
     $script:testNum = 0
+    $script:credentials = Import-Clixml -Path (Join-Path -Path $PSScriptRoot -ChildPath '..\.cenvironment' -Resolve)
 
     function Assert-NoTestEnvironmentVariableAt( $Scope )
     {
@@ -39,11 +48,16 @@ AfterAll {
         Where-Object { $_ -like 'CARBON_REMOVEENVVAR_TEST_*' } |
         Select-Object -Unique |
         ForEach-Object {
-            Remove-CEnvironmentVariable -Name $_ -ForProcess -ForUser -ForComputer
+            $forComputerArg = @{}
+            if (Test-TCRunAsElevated)
+            {
+                $forComputerArg['ForComputer'] = $true
+            }
+            Remove-CEnvVariable -Name $_ -ForProcess -ForUser @forComputerArg
         }
 }
 
-Describe 'Remove-CEnvironmentVariable' {
+Describe 'Remove-CEnvVariable' {
     BeforeEach {
         while ($true)
         {
@@ -61,27 +75,28 @@ Describe 'Remove-CEnvironmentVariable' {
         $Global:Error.Clear()
     }
 
-    It 'removes computer-level variable' {
+    It 'removes computer-level variable' -Skip:(-not (Test-TCRunAsElevated)) {
         Set-TestEnvironmentVariable 'Machine'
-        Remove-CEnvironmentVariable -Name $script:varName -ForComputer
+        Remove-CEnvVariable -Name $script:varName -ForComputer
         Assert-NoTestEnvironmentVariableAt -Scope Machine
     }
 
     It 'removes user-level variable' {
         Set-TestEnvironmentVariable 'User'
-        Remove-CEnvironmentVariable -Name $script:varName -ForUser
+        Remove-CEnvVariable -Name $script:varName -ForUser
         Assert-NoTestEnvironmentVariableAt -Scope User
     }
 
     It 'removes process-level variable' {
         Set-TestEnvironmentVariable 'Process'
-        Remove-CEnvironmentVariable -Name $script:varName -ForProcess
+        Remove-CEnvVariable -Name $script:varName -ForProcess
         Assert-NoTestEnvironmentVariableAt -Scope Process
     }
 
     Context '<_> scope' -ForEach 'Computer','User','Process' {
         $scope = $_
-        It 'removes variable with the Force' -ForEach $scope {
+        $skip = $scope -eq 'Computer' -and -not (Test-TCRunAsElevated)
+        It 'removes variable with the Force' -ForEach $scope -Skip:$skip {
             $scope = $_
             $setScope = $scope
             if( $scope -eq 'Computer' )
@@ -92,21 +107,21 @@ Describe 'Remove-CEnvironmentVariable' {
             $scopeParam = @{
                                 ('For{0}' -f $scope) = $true
                         }
-            Remove-CEnvironmentVariable -Name $script:varName @scopeParam -Force
+            Remove-CEnvVariable -Name $script:varName @scopeParam -Force
             Assert-NoTestEnvironmentVariableAt -Scope $setScope
             Test-Path -Path ('env:{0}' -f $script:varName) | Should -BeFalse
         }
     }
 
     It 'ignores non-existent variable' {
-        Remove-CEnvironmentVariable -Name "IDoNotExist" -ForComputer
+        Remove-CEnvVariable -Name "IDoNotExist" -ForProcess
         $Global:Error | Should -BeNullOrEmpty
     }
 
     It 'supports WhatIf' {
         $envVarValue = Set-TestEnvironmentVariable -Scope Process
 
-        Remove-CEnvironmentVariable -Name $script:varName -ForProcess -WhatIf
+        Remove-CEnvVariable -Name $script:varName -ForProcess -WhatIf
 
         $actualValue = [Environment]::GetEnvironmentVariable($script:varName, 'Process')
         $actualValue | Should -Not -BeNullOrEmpty
@@ -115,27 +130,33 @@ Describe 'Remove-CEnvironmentVariable' {
 
     It 'removes from all scopes at once' {
         $value = [Guid]::NewGuid().ToString()
-        Set-CEnvironmentVariable -Name $script:varName -Value $value -ForProcess -ForUser -ForComputer
-        Remove-CEnvironmentVariable -Name $script:varName -ForProcess -ForUser -ForComputer
+        $forComputerScopeArg = @{}
+        if (Test-TCRunAsElevated)
+        {
+            $forComputerScopeArg['ForComputer'] = $true
+        }
+
+        Set-CEnvVariable -Name $script:varName -Value $value -ForProcess -ForUser @forComputerScopeArg
+        Remove-CEnvVariable -Name $script:varName -ForProcess -ForUser @forComputerScopeArg
         Assert-NoTestEnvironmentVariableAt -Scope Machine
         Assert-NoTestEnvironmentVariableAt -Scope User
         Assert-NoTestEnvironmentVariableAt -Scope Process
     }
 
     It 'requires at least one scope' {
-        Remove-CEnvironmentVariable -Name $script:varName -ErrorAction SilentlyContinue
+        Remove-CEnvVariable -Name $script:varName -ErrorAction SilentlyContinue
         $Global:Error | Should -Match 'target not specified'
     }
 
     It 'removes variable for another user' {
         $name = [Guid]::NewGuid().ToString()
         $value = [Guid]::NewGuid().ToString()
-        Set-CEnvironmentVariable -Name $name -Value $value -ForUser -Credential $CarbonTestUser
-        Remove-CEnvironmentVariable -Name $name -ForUser -Credential $CarbonTestUser
+        Set-CEnvVariable -Name $name -Value $value -ForUser -Credential $script:credentials
+        Remove-CEnvVariable -Name $name -ForUser -Credential $script:credentials
         $actualValue = $value
         $job = Start-Job -ScriptBlock {
             Get-Item -Path ('env:{0}' -f $using:name) -ErrorAction Ignore
-        } -Credential $CarbonTestUser
+        } -Credential $script:credentials
         $actualValue = $job | Wait-Job | Receive-Job
         $job | Remove-Job -Force -ErrorAction Ignore
         $actualValue | Should -BeNullOrEmpty